Edited Python File
Edited Python file, specifically the show_schedule and add_schedule functions, so they are able to add the daily schedule from each dictionary to the .txt file and return the full  weekly schedule. I also edited the main function to run smoothly with my changes to the show_schedule and add_schedule functions.